python-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pillow==10.4.0
ultralytics==8.3.0
pydantic==2.9.0
python-multipart==0.0.12
